// src/api/generated/model/createRoomOrdContractRequestDtoActionType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */

/**
 * Тип действия
 */
export type CreateRoomOrdContractRequestDtoActionType = typeof CreateRoomOrdContractRequestDtoActionType[keyof typeof CreateRoomOrdContractRequestDtoActionType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateRoomOrdContractRequestDtoActionType = {
  distribution: 'distribution',
  conclude: 'conclude',
  commercial: 'commercial',
  other: 'other',
} as const;