// src/api/generated/model/createRoomOrdContractRequestDtoSubjectType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */

/**
 * Тип предмета договора
 */
export type CreateRoomOrdContractRequestDtoSubjectType = typeof CreateRoomOrdContractRequestDtoSubjectType[keyof typeof CreateRoomOrdContractRequestDtoSubjectType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateRoomOrdContractRequestDtoSubjectType = {
  representation: 'representation',
  org_distribution: 'org_distribution',
  mediation: 'mediation',
  distribution: 'distribution',
  other: 'other',
} as const;