// src/api/generated/model/createRoomOrdContractRequestDtoType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */

/**
 * Тип договора
 */
export type CreateRoomOrdContractRequestDtoType = typeof CreateRoomOrdContractRequestDtoType[keyof typeof CreateRoomOrdContractRequestDtoType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateRoomOrdContractRequestDtoType = {
  service: 'service',
  mediation: 'mediation',
  additional: 'additional',
} as const;